Split moorhen-wrappers.cc into manageable files
Split monolithic moorhen-wrappers.cc (~2968 lines) into:
- moorhen-wrappers-helpers.h: shared includes, structs, helper functions,
  molecules_container_js class definition
- moorhen-wrappers.cc: molecules_container_t and molecules_container_js
  embind bindings (the main API surface)
- moorhen-types-wrappers.cc: all supporting types (value_objects,
  register_vectors, clipper/mmdb classes, coot utility types, pairs,
  value_arrays, free functions)

Also add inline to validate() in privateer-wrappers.h to fix duplicate
symbol when the header is included from multiple translation units.

All 16 gemmi tests pass.
